--- a/packages/benchmarks/src/benchmarks/index.ts
+++ b/packages/benchmarks/src/benchmarks/index.ts
@@ -1,2 +1,3 @@
 //import "./map-vs-btree.ts";
 import "./blinkdb-vs-lokijs.ts";
+//import "./blinkdb.ts";