depends on #3 for additional benchmarking (therefore this PR contains the other changes as well until it's merged).

I have changed `limitItems` to need way less array operations when limiting/paginating a large item data set.

I also fixed a bug with `many` which didn't return any items if filter options are defined but don't contain a `where` (e.g. when paginating the whole data set).

## many query with limit:
### Before
```
  blinkdb   took 0.41430ms, high 16.08050ms, low 0.40960ms
```
### After
```
  blinkdb   took 0.19070ms, high 2.21160ms, low 0.17670ms
```