@froehlichA 2022-09-13
Performance increases on `sort` and `limit` are always welcome, primarily because those are the least optimized parts of the query engine ^^