@froehlichA 2022-09-13
As far as I can tell, the first commit in this PR is from #3, and the other two look fine as is.
I'll merge this after #3 is done.